requirements.txt
requests>=2.31.0
pyyaml>=6.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0